Add Google Drive

+++ src/main/java/ubicrypt/core/provider/gdrive/GDriveProvider.java
@@ -3,6 +3,8 @@ package ubicrypt.core.provider.gdrive;
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow;
 import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets;
+import com.google.api.client.googleapis.json.GoogleJsonResponseException;
+import com.google.api.client.http.HttpTransport;
 import com.google.api.client.http.InputStreamContent;
 import com.google.api.client.http.apache.ApacheHttpTransport;
 import com.google.api.client.util.store.DataStore;
@@ -10,74 +12,100 @@ import com.google.api.client.util.store.DataStoreFactory;
 import com.google.api.services.drive.Drive;
 import com.google.api.services.drive.DriveRequestInitializer;
 import com.google.api.services.drive.DriveScopes;
+import com.google.api.services.drive.model.About;
 import com.google.api.services.drive.model.File;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Serializable;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicReference;
 
 import rx.Observable;
+import ubicrypt.core.exp.NotFoundException;
 import ubicrypt.core.provider.ProviderStatus;
 import ubicrypt.core.provider.UbiProvider;
 
 import static com.google.api.client.json.jackson2.JacksonFactory.getDefaultInstance;
 import static java.util.Arrays.asList;
+import static java.util.Collections.singletonList;
+import static org.apache.commons.lang3.StringUtils.isEmpty;
+import static org.slf4j.LoggerFactory.getLogger;
 
 public class GDriveProvider extends UbiProvider implements DataStoreFactory {
+    private static final Logger log = getLogger(GDriveProvider.class);
+    public static String credentialId = "id";
 
     @JsonInclude
-    private Set<GDriveConf> confs = ConcurrentHashMap.newKeySet();
+    private GDriveConf conf;
     private volatile Drive drive;
-    private volatile AtomicReference<Credential> credential = new AtomicReference<>();
-    private volatile String folderId;
+    private volatile HttpTransport http = new ApacheHttpTransport();
 
     public GDriveProvider(GDriveConf conf) {
-        confs.add(conf);
+        this.conf = conf;
+    }
+
+    public GDriveProvider() {
     }
 
     @Override
     public Observable<ProviderStatus> init(long userId) {
         return Observable.create(subscriber -> {
+            log.debug("init provider userId:{}", userId);
             try {
-                ApacheHttpTransport http = new ApacheHttpTransport();
                 GoogleAuthorizationCodeFlow flow =
                         new GoogleAuthorizationCodeFlow.Builder(
                                 http, getDefaultInstance(), appCredentials(), asList(DriveScopes.DRIVE_FILE))
                                 .setDataStoreFactory(this)
                                 .setAccessType("offline")
                                 .build();
-                confs.stream().forEach(c -> c.setDataStoreFactory(this));
-                credential.set(flow.loadCredential(String.valueOf(userId)));
-                if (credential.get() == null || credential.get().getRefreshToken() == null) {
-                    subscriber.onNext(ProviderStatus.unavailable);
+                conf.setDataStoreFactory(this);
+                Credential credential = flow.loadCredential(credentialId);
+                if (credential == null || credential.getRefreshToken() == null) {
+                    subscriber.onNext(ProviderStatus.unauthorized);
                     return;
                 }
-
-                drive = new Drive.Builder(http, getDefaultInstance(), credential.get())
+                drive = new Drive.Builder(http, getDefaultInstance(), credential)
                         .setDriveRequestInitializer(new DriveRequestInitializer())
                         .setApplicationName("UbiCrypt").build();
+                String email = null;
+                try {
+                    email = drive.about().get().setFields("user, storageQuota").execute().getUser().getEmailAddress();
+                } catch (GoogleJsonResponseException e) {
+                    if (e.getDetails().getCode() == 403) {
+                        drive = null;
+                        log.info("gdrive unauthorized:{}", e.getDetails().getMessage());
+                        subscriber.onNext(ProviderStatus.unauthorized);
+                        subscriber.onCompleted();
+                        return;
+                    }
+                }
+                if (isEmpty(conf.getEmail())) {
+                    conf.setEmail(email);
+                } else if (!StringUtils.equals(conf.getEmail(), email)) {
+                    log.warn("this provider's gdrive email is not the same of the current google account");
+                    subscriber.onNext(ProviderStatus.unavailable);
+                }
                 File folder = new File();
-                folder.setName("ubicrypt");
+                folder.setName(conf.getFolderName());
                 folder.setMimeType("application/vnd.google-apps.folder");
-                final DataStore<Serializable> dataStore = getDataStore(String.valueOf(userId));
-                if (!dataStore.containsKey("folder")) {
+                if (isEmpty(conf.getFolderId())) {
+                    log.debug("create new ubicrypt folder");
                     File createdFolder = drive.files().create(folder).execute();
-                    folderId = createdFolder.getId();
-                    dataStore.set("folder", createdFolder.getId());
+                    conf.setFolderId(createdFolder.getId());
                 } else {
-                    folderId = (String) dataStore.get("folder");
+                    String folderId = conf.getFolderId();
                     try {
                         drive.files().get(folderId).execute();
+                        log.debug("ubicrypt folder already present");
                     } catch (Exception e) {
+                        log.debug("create new ubicrypt folder");
                         File createdFolder = drive.files().create(folder).execute();
-                        folderId = createdFolder.getId();
-                        dataStore.set("folder", createdFolder.getId());
+                        conf.setFolderId(createdFolder.getId());
                     }
                 }
                 subscriber.onNext(ProviderStatus.initialized);
@@ -89,10 +117,19 @@ public class GDriveProvider extends UbiProvider implements DataStoreFactory {
     }
 
     @Override
+    public String code() {
+        return "gdrive";
+    }
+
+    @Override
     public Observable<String> post(InputStream is) {
         return Observable.create(subscriber -> {
+            if (drive == null || conf == null || conf.getFolderId() == null) {
+                subscriber.onError(new RuntimeException("gdrive not initialized"));
+                return;
+            }
             File file = new File();
-            file.getParents().add(folderId);
+            file.setParents(singletonList(conf.getFolderId()));
             try {
                 String id = drive.files().create(file, new InputStreamContent(null, is))
                         .execute()
@@ -110,6 +147,10 @@ public class GDriveProvider extends UbiProvider implements DataStoreFactory {
     public Observable<Boolean> delete(String pid) {
         return Observable.create(subscriber -> {
             try {
+                if (drive == null) {
+                    subscriber.onError(new RuntimeException("gdrive not initialized"));
+                    return;
+                }
                 drive.files().delete(pid);
                 subscriber.onNext(true);
                 subscriber.onCompleted();
@@ -122,10 +163,13 @@ public class GDriveProvider extends UbiProvider implements DataStoreFactory {
     @Override
     public Observable<Boolean> put(String pid, InputStream is) {
         return Observable.create(subscriber -> {
+            if (drive == null) {
+                subscriber.onError(new RuntimeException("gdrive not initialized"));
+                return;
+            }
             File file = new File();
-            file.setName(pid);
             try {
-                drive.files().create(file, new InputStreamContent(null, is))
+                drive.files().update(pid, file, new InputStreamContent(null, is))
                         .execute();
                 subscriber.onNext(true);
                 subscriber.onCompleted();
@@ -140,9 +184,17 @@ public class GDriveProvider extends UbiProvider implements DataStoreFactory {
     @Override
     public Observable<InputStream> get(String pid) {
         return Observable.create(subscriber -> {
+            if (drive == null) {
+                subscriber.onError(new RuntimeException("gdrive not initialized"));
+                return;
+            }
             try {
                 subscriber.onNext(drive.files().get(pid).executeMediaAsInputStream());
                 subscriber.onCompleted();
+            } catch (GoogleJsonResponseException e) {
+                if (e.getDetails().getCode() == 404) {
+                    subscriber.onError(new NotFoundException(pid));
+                }
             } catch (IOException e) {
                 subscriber.onError(e);
             }
@@ -152,23 +204,55 @@ public class GDriveProvider extends UbiProvider implements DataStoreFactory {
 
     @Override
     public String providerId() {
-        return null;
+        return "gdrive://" + conf.getEmail();
     }
 
     private static GoogleClientSecrets appCredentials() throws IOException {
         return GoogleClientSecrets.load(getDefaultInstance(),
-                new InputStreamReader(DriveSample.class.getResourceAsStream("/google.json")));
+                new InputStreamReader(GDriveProvider.class.getResourceAsStream("/google.json")));
     }
 
     @Override
     public <V extends Serializable> DataStore<V> getDataStore(String id) throws IOException {
-        return confs.stream()
-                .filter(conf -> conf.getId().equals(id))
-                .findFirst()
-                .orElseGet(() -> {
-                    GDriveConf conf = new GDriveConf(id, this);
-                    confs.add(conf);
-                    return conf;
-                });
+        return conf;
+    }
+
+    public GDriveConf getConf() {
+        return conf;
+    }
+
+    public void setConf(GDriveConf conf) {
+        this.conf = conf;
+    }
+
+    @Override
+    public String toString() {
+        return providerId();
+    }
+
+    @Override
+    public Observable<Long> availableSpace() {
+        return Observable.create(subscriber -> {
+            if (drive == null) {
+                subscriber.onError(new RuntimeException("gdrive not initialized"));
+                return;
+            }
+            final About storageQuota;
+            try {
+                storageQuota = drive.about().get().setFields("storageQuota").execute();
+            } catch (IOException e) {
+                subscriber.onError(e);
+                return;
+            }
+            Long limit = storageQuota.getStorageQuota().getLimit();
+            if (limit == null) {
+                subscriber.onNext(Long.MAX_VALUE);
+                subscriber.onCompleted();
+                return;
+            }
+            subscriber.onNext(limit - storageQuota.getStorageQuota().getUsage());
+            subscriber.onCompleted();
+
+        });
     }
 }
